+++ tests/test_app/tests/test_api.py
@@ -5,6 +5,7 @@ from django.urls import reverse
 from django.utils.http import urlencode
 from rest_framework import status
 from rest_framework.test import APITestCase
+
 from test_app.models import TestLimitedModel, TestModel